schooldriver-modern/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
drf-spectacular==0.27.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
whitenoise==6.6.0
